requirements.txt
azure-functions==1.*
msal==1.*
requests==2.*
pytz
python-dateutil
